backend-python/ct_client/pyproject.toml
[tool.poetry]
name = "clinical-trials-gov-rest-api-client"
version = "2.0.3"
description = "A client library for accessing ClinicalTrials.gov REST API"
authors = []
readme = "README.md"
packages = [
    {include = "clinical_trials_gov_rest_api_client"},
]
include = ["CHANGELOG.md", "clinical_trials_gov_rest_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
